% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkBlockDownload.R
\name{bulkBlockDownload}
\alias{bulkBlockDownload}
\title{Bulk Block Download}
\usage{
bulkBlockDownload(start_block, end_block, data_dir, chunk_size = 50000,
  parallel = TRUE, cores = 3)
}
\arguments{
\item{start_block}{numeric, Starting block number.}

\item{end_block}{numeric, Ending block number.}

\item{data_dir}{path, Desired folder for storing analysis table in.}

\item{chunk_size}{numeric, Chunk size to Download.}

\item{parallel}{logical, Should plyr operate in parallel.}

\item{cores}{numeric, number of cores to be intialised. Suggested n - 1, 
where n is localhosts number of cores.}
}
\value{
saves .RData file within \code{ethr_blocks} folder in the 
  \code{data_dir}.
}
\description{
Breaks a large block range up and manages the download size. All downloaded 
blocks will be stored in a \code{ethr_blocks} folder in the \code{data_dir}. 
Uses doParallel package to create a multithread workgroup to increase 
download speed.
}

