% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkBlockDownload.R
\name{combineDownloadedBlocks}
\alias{combineDownloadedBlocks}
\title{Combine Downloaded Transaction Chunks}
\usage{
combineDownloadedBlocks(data_dir)
}
\arguments{
\item{data_dir}{path, Desired folder for storing analysis table in.}
}
\value{
a tbl_df of the combined download chunks stored in \code{ethr_blocks}.
}
\description{
Reads all .RData files within \code{ethr_blocks} and combines them into a
single data.frame for analysis.
}

