% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{decHex}
\alias{decHex}
\title{Convert decimal to hex value}
\usage{
decHex(dec)
}
\arguments{
\item{dec}{numeric or character, an integer value formatted as character if
over 32bit max (2147483647).}
}
\value{
character, hex equivalent, '0x' prefixed.
}
\description{
Convert decimal to hex value
}

