% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_accounts}
\alias{eth_accounts}
\title{Eth accounts}
\usage{
eth_accounts(rpc_address = "http://localhost:8545")
}
\arguments{
\item{rpc_address}{The address of the RPC API.}
}
\value{
20 Bytes - addresses owned by the client.
}
\description{
Returns a list of addresses owned by client.
}

