% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_blockNumber}
\alias{eth_blockNumber}
\title{Eth blockNumber}
\usage{
eth_blockNumber(rpc_address = "http://localhost:8545")
}
\arguments{
\item{rpc_address}{The address of the RPC API.}
}
\value{
hex value of the current block number the client is on.
}
\description{
Returns the number of most recent block.
}

