% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_gasPrice}
\alias{eth_gasPrice}
\title{Eth gasPrice}
\usage{
eth_gasPrice(rpc_address = "http://localhost:8545")
}
\arguments{
\item{rpc_address}{The address of the RPC API.}
}
\value{
hex value of the current gas price in wei.
}
\description{
Returns the current price per gas in wei.
}

