% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getBalance}
\alias{eth_getBalance}
\title{Eth getBalance}
\usage{
eth_getBalance(address, block_number, rpc_address = "http://localhost:8545")
}
\arguments{
\item{address}{20 Bytes - address to check for balance.}

\item{block_number}{hex value of block number, or the string "latest", "earliest" or "pending".}

\item{rpc_address}{The address of the RPC API.}
}
\value{
hex value of the current balance in wei.
}
\description{
Returns the balance of the account of given address.
}

