% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getBlockTransactionCountByHash}
\alias{eth_getBlockTransactionCountByHash}
\title{eth_getBlockTransactionCountByHash}
\usage{
eth_getBlockTransactionCountByHash(block_hash,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{block_hash}{32 Bytes - hash of a block.}

\item{rpc_address}{The address of the RPC API.}
}
\value{
hex value of the number of transactions in this block.
}
\description{
Returns the number of transactions in a block from a block matching the given block hash.
}

