% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getBlockTransactionCountByNumber}
\alias{eth_getBlockTransactionCountByNumber}
\title{Eth getBlockTransactionCountByNumber}
\usage{
eth_getBlockTransactionCountByNumber(block_number,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{block_number}{hex value of a block number, or the string "earliest", "latest" or "pending".}

\item{rpc_address}{The address of the RPC API.}
}
\value{
QUANTITY - integer of the number of transactions in this block, hex format
}
\description{
Returns the number of transactions in a block from a block matching the given block number.
}

