% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getCode}
\alias{eth_getCode}
\title{Eth getCode}
\usage{
eth_getCode(address, block_number, rpc_address = "http://localhost:8545")
}
\arguments{
\item{rpc_address}{The address of the RPC API.}

\item{20}{Bytes - address.}

\item{hex}{value of block number, or the string "latest", "earliest" or "pending".}
}
\value{
the code from the given address.
}
\description{
Returns code at a given address.
}

