% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getStorageAt}
\alias{eth_getStorageAt}
\title{Eth getStorageAt}
\usage{
eth_getStorageAt(address, position_number, block_number,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{address}{20 Bytes - address of the storage.}

\item{position_number}{hex value of the position in the storage.}

\item{block_number}{hex value of block number, or the string "latest", "earliest" or "pending".}

\item{rpc_address}{The address of the RPC API.}
}
\value{
the value at this storage position.
}
\description{
Returns the value from a storage position at a given address.
}

