% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getTransactionByBlockHashAndIndex}
\alias{eth_getTransactionByBlockHashAndIndex}
\title{Eth getTransactionByBlockHashAndIndex}
\usage{
eth_getTransactionByBlockHashAndIndex(block_hash, index_number = "0x0",
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{block_hash}{32 Bytes - hash of a block.}

\item{index_number}{integer of the transaction index position.}

\item{rpc_address}{The address of the RPC API.}
}
\value{
List of transaction properties, or null when no transaction found
}
\description{
Returns information about a transaction by block hash and transaction index position.
}

