% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getTransactionByBlockNumberAndIndex}
\alias{eth_getTransactionByBlockNumberAndIndex}
\title{Eth getTransactionByBlockNumberAndIndex}
\usage{
eth_getTransactionByBlockNumberAndIndex(block_number, index_number = "0x0",
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{block_number}{a hex block number, or the string "earliest", "latest" or "pending".}

\item{index_number}{integer of the transaction index position.}

\item{rpc_address}{The address of the RPC API.}
}
\value{
List of transaction properties, or null when no transaction found

hash: DATA, 32 Bytes - hash of the transaction.

nonce: numeric - the number of transactions made by the sender prior to this one.

blockHash: DATA, 32 Bytes - hash of the block where this transaction was in. null when its pending.

blockNumber: QUANTITY - block number where this transaction was in. null when its pending.

transactionIndex: QUANTITY - integer of the transactions index position in the block. null when its pending.

from: DATA, 20 Bytes - address of the sender.

to: DATA, 20 Bytes - address of the receiver. null when its a contract creation transaction.

value: QUANTITY - value transferred in Wei.

gasPrice: QUANTITY - gas price provided by the sender in Wei.

gas: QUANTITY - gas provided by the sender.

input: DATA - the data send along with the transaction.
}
\description{
Returns information about a transaction by block number and transaction index position.
}

