% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getTransactionByHash}
\alias{eth_getTransactionByHash}
\title{eth_getTransactionByHash}
\usage{
eth_getTransactionByHash(transaction_hash,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{transaction_hash}{32 Bytes - hash of a transaction.}

\item{rpc_address}{The address of the RPC API.}
}
\value{
List of transaction properties, or null when no transaction found.

hash: DATA, 32 Bytes - hash of the transaction.

nonce: numeric - the number of transactions made by the sender prior to this one.

blockHash: DATA, 32 Bytes - hash of the block where this transaction was in. null when its pending.

blockNumber: QUANTITY - block number where this transaction was in. null when its pending.

transactionIndex: QUANTITY - integer of the transactions index position in the block. null when its pending.

from: DATA, 20 Bytes - address of the sender.

to: DATA, 20 Bytes - address of the receiver. null when its a contract creation transaction.

value: QUANTITY - value transferred in Wei.

gasPrice: QUANTITY - gas price provided by the sender in Wei.

gas: QUANTITY - gas provided by the sender.

input: DATA - the data send along with the transaction.
}
\description{
Returns the information about a transaction requested by transaction hash.
}

