% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getTransactionCount}
\alias{eth_getTransactionCount}
\title{eth_getTransactionCount}
\usage{
eth_getTransactionCount(address, block_number,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{address}{20 Bytes - address.}

\item{block_number}{hex value of block number, or the string "latest", "earliest" or "pending".}

\item{rpc_address}{The address of the RPC API.}
}
\value{
hex of the number of transactions send from this address.
}
\description{
Returns the number of transactions sent from an address - state at given block number.
}

