% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethRPCFunctions.R
\name{eth_getTransactionReceipt}
\alias{eth_getTransactionReceipt}
\title{eth_getTransactionReceipt}
\usage{
eth_getTransactionReceipt(transaction_hash,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{transaction_hash}{32 Bytes - hash of a transaction.}

\item{rpc_address}{The address of the RPC API.}
}
\value{
transationHash: DATA, 32 Bytes - hash of the transaction.

transactionIndex: QUANTITY - integer of the transactions index position in the block. null when its pending.

blockHash: DATA, 32 Bytes - hash of the block where this transaction was in. null when its pending.

blockNumber: QUANTITY - block number where this transaction was in. null when its pending.

cumulativeGasUsed: QUANTITY - The total amount of gas used when this transaction was executed in the block.

gasUsed: QUANTITY - The amount of gas used by this specific transaction alone.

vcontractAddress: DATA, 20 Bytes - The contract address created, if the transaction was a contract creation, otherwise null.

logs: Array - Array of log objects, which this transaction generated.
}
\description{
Returns the receipt of a transaction by transaction hash.
}

