% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{ethCon}
\alias{ethCon}
\title{Convert Ethereum Denominations}
\usage{
ethCon(x, from, to)
}
\arguments{
\item{x}{numeric, number to be converted}

\item{from}{character, 2 letter short notation to convert from (see details)}

\item{to}{character, 2 letter short notation to convert to (see details)}
}
\value{
mpfr converted value (see details)
}
\description{
Convert Ethereum Denominations
}
\details{
This function returns a Multiple Precision Floating-Point Reliable
number (implemented by the Rmpfr package). This is to allow for the conersion
of very high values of Wei into other forms. This can then be converted to
numeric, integer or double as required at the expense of accuracy. WARNING do
not use this in calculating values to transact actual ether as the precision is not
guaranteed currently. This is intended for indicative conversions when parsing
data from the Ethereum blockchain. \cr
\cr
Denomination short codes:\cr
wei = we, Kwei = kw, Mwei = mw,  Gwei = gw \cr
szabo = sz, finney = fi, ether = et \cr
Kether = ke, Mether = me,Gether = ge, Tether = te \cr
}
\examples{
ethCon(15000000000000000, "we", "et")
ethCon(1000000000, "et", "me")
}

