% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockHeaders.R
\name{getBlockHeaders}
\alias{getBlockHeaders}
\title{Block Headers}
\usage{
getBlockHeaders(n_blocks = NULL, start_block = NULL, end_block = NULL,
  parallel = FALSE, rpc_address = "http://localhost:8545")
}
\arguments{
\item{n_blocks}{numeric, Number of blocks to extract back from the latest block.}

\item{start_block}{numeric, Starting block number for extraction.}

\item{end_block}{numeric, Ending block for extraction, inclusive.}

\item{parallel}{boolean, invokes a parallelised plyr loop}

\item{rpc_address}{string, The address of the RPC API.

To initialise parallel workers:
workers <-makeCluster(3)
doParallel::registerDoParallel(workers)}
}
\value{
data.frame Block header data of requested blocks.
}
\description{
Queries the blockchain and returns a data.frame of the block headers for either a given
number of blocks or a specified block range of blocks.
}
\examples{
getBlockHeaders(100)
getBlockHeaders(start_block = 951460, end_block = 951560)
}

