% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockTransactions.R
\name{getBlockTransactions}
\alias{getBlockTransactions}
\title{Block Transaction Data}
\usage{
getBlockTransactions(n_blocks = NULL, start_block = NULL,
  end_block = NULL, parallel = FALSE,
  rpc_address = "http://localhost:8545")
}
\arguments{
\item{n_blocks}{numeric, Number of blocks to extract back from the latest block.}

\item{start_block}{numeric, Starting block number for extraction.}

\item{end_block}{numeric, Ending block for extraction, inclusive.}

\item{rpc_address}{character, The address of the RPC API.}
}
\value{
data.frame Transation data of requested blocks.
}
\description{
Queries the blockchain and returns a data.frame of transactions for either a given
number of blocks or a specified block range of blocks.
}
\examples{
getBlockTransactions(100)
getBlockTransactions(start_block = 951460, end_block = 951560)
}

