% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{hexDec}
\alias{hexDec}
\title{Convert hex to decimal value}
\usage{
hexDec(hex)
}
\arguments{
\item{hex}{character, hex value, '0x' prefixed.}
}
\value{
numeric, decimal equivalent
}
\description{
Convert hex to decimal value
}

