% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{hexLeading}
\alias{hexLeading}
\title{Swap hex value types}
\usage{
hexLeading(hex)
}
\arguments{
\item{hex}{character, a hex value}
}
\value{
character,
}
\description{
Alternate between '0x' and non-'0x' prefixed hex values
}

