% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodePlot.R
\name{nodesPlot}
\alias{nodesPlot}
\title{nodesPlots}
\usage{
nodesPlot(table = transactions, degree_thres = 10, save_plot = FALSE)
}
\arguments{
\item{table, }{dataframe containing transactions, eg output from getBlockTransactions}

\item{degree_thres}{numeric, number of transactions, used to select nodes with high in- or out-degrees.}

\item{save_plot, }{save the plot to disk. Default is FALSE.}
}
\value{
plot of the nodes and transaction for the input data.
}
\description{
Plots the nodes/accounts and transaction between said accounts for input transactions data.
The nodes with high in and out degree activity are coloured. The threshold value is 10. 
Large transactions (greater than 3 standard deviation from the mean transaction value) are shown with black arrows.
}
\examples{
nodesPlot(transactions)
nodesPlot(transaction,degree_thres=50,save_plot=TRUE)
}

